+++ extensions/my-daily-log/src/factories/checkLogReminderUseCaseFactory.tsx
@@ -1,0 +1,6 @@
+import { CheckLogReminderUseCase } from "../domain/dailyLog/useCases/CheckLogReminderUseCase";
+import { makeDailyLogRepository } from "./makeDailyLogRepository";
+
+export function makeCheckLogReminderUseCase(): CheckLogReminderUseCase {
+  return new CheckLogReminderUseCase(makeDailyLogRepository());
+}